EQ2-168 made change base on pull request

+++ src/main/java/projet/projetstage02/dto/notification/CompanyNotificationDTO.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class CompanyNotificationDTO {
-    private int nbOffers;
-    private int nbContracts;
+    private long nbOffers;
+    private long nbContracts;
 }
